optimalize siege processing using queues

+++ src/main/java/com/flansmod/warforge/common/WarForgeConfig.java
@@ -58,7 +58,7 @@ public class WarForgeConfig {
     public static int SIEGE_SWING_PER_ATTACKER_FLAG = 1;
     public static int SIEGE_DIFFICULTY_PER_DEFENDER_FLAG = 3;
     public static boolean SIEGE_CAPTURE = false;
-    public static float SIEGE_COOLDOWN_FAIL = 30f; // actually essentially in ms
+    public static int SIEGE_COOLDOWN_FAIL = 30; // in minutes;
     public static float FLAG_COOLDOWN = 1f; // In minutes
     public static int MAX_SIEGES = 3;
     public static int ATTACKER_DESERTION_TIMER = 180; // in seconds
